requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
altgraph==0.17.5
    # via
    #   macholib
    #   pyinstaller
certifi==2026.1.4
    # via
    #   -c requirements.txt
    #   requests
charset-normalizer==3.4.4
    # via
    #   -c requirements.txt
    #   requests
gitdb==4.0.12
    # via
    #   -c requirements.txt
    #   gitpython
gitpython==3.1.46
    # via
    #   -c requirements.txt
    #   -r requirements.in
idna==3.11
    # via
    #   -c requirements.txt
    #   requests
iniconfig==2.3.0
    # via pytest
macholib==1.16.4
    # via pyinstaller
markdown-it-py==4.0.0
    # via
    #   -c requirements.txt
    #   rich
markupsafe==3.0.3
    # via werkzeug
mdurl==0.1.2
    # via
    #   -c requirements.txt
    #   markdown-it-py
packaging==26.0
    # via
    #   pyinstaller
    #   pyinstaller-hooks-contrib
    #   pytest
pluggy==1.6.0
    # via pytest
pygments==2.19.2
    # via
    #   -c requirements.txt
    #   pytest
    #   rich
pyinstaller==6.18.0
    # via -r requirements-dev.in
pyinstaller-hooks-contrib==2026.0
    # via pyinstaller
pytest==9.0.2
    # via -r requirements-dev.in
pytest-httpserver==1.1.5
    # via -r requirements-dev.in
requests==2.32.5
    # via
    #   -c requirements.txt
    #   -r requirements.in
rich==14.3.2
    # via
    #   -c requirements.txt
    #   -r requirements.in
ruff==0.15.0
    # via -r requirements-dev.in
smmap==5.0.2
    # via
    #   -c requirements.txt
    #   gitdb
urllib3==2.6.3
    # via
    #   -c requirements.txt
    #   -r requirements.in
    #   requests
werkzeug==3.1.5
    # via pytest-httpserver

# The following packages are considered to be unsafe in a requirements file:
# setuptools
